kiosk-app/requirements.txt
PyQt5==5.15.4
paho-mqtt==1.5.1
